+++ R/GSManager.R
@@ -4,8 +4,10 @@
 #' @importFrom R6 R6Class
 #' @importFrom openssl base64_encode
 #' @import httr
-#' @import XML
+#' @import xml2
 #' @import keyring
+#' @import cli
+#' @import magrittr
 #' @importFrom readr read_csv
 #' @importFrom readr write_csv
 #' @export
@@ -17,50 +19,6 @@
 #' @examples
 #' \dontrun{
 #'    GSManager$new("http://localhost:8080/geoserver", "admin", "geoserver")
-#' }
-#'
-#' @section Methods:
-#' \describe{
-#'  \item{\code{new(url, user, pwd, logger, keyring_backend)}}{
-#'    
-#'  }
-#'  \item{\code{logger(type, text)}}{
-#'    Basic logger to report geosapi logs. Used internally
-#'  }
-#'  \item{\code{INFO(text)}}{
-#'    Logger to report information. Used internally
-#'  }
-#'  \item{\code{WARN(text)}}{
-#'    Logger to report warnings. Used internally
-#'  }
-#'  \item{\code{ERROR(text)}}{
-#'    Logger to report errors. Used internally
-#'  }
-#'  \item{\code{getUrl()}}{
-#'    Get the authentication URL
-#'  }
-#'  \item{\code{connect()}}{
-#'    This methods attempts a connection to GeoServer REST API. User internally
-#'    during initialization of \code{GSManager}.
-#'  }
-#'  \item{\code{reload()}}{
-#'    Reloads the GeoServer catalog.
-#'  }
-#'  \item{\code{getSystemStatus()}}{
-#'    Get system status
-#'  }
-#'  \item{\code{getClassName()}}{
-#'    Retrieves the name of the class instance
-#'  }
-#'  \item{\code{getWorkspaceManager()}}{
-#'    Retrieves an instance of workspace manager
-#'  }
-#'  \item{\code{getNamespaceManager()}}{
-#'    Retrieves an instance of namespace manager
-#'  }
-#'  \item{\code{getDataStoreManager()}}{
-#'    Retrieves an instance of datastore manager
-#'  }
 #' }
 #' 
 #' @author Emmanuel Blondel <emmanuel.blondel1@@gmail.com>
@@ -211,20 +169,25 @@ GSManager <- R6Class("GSManager",
       )
       if(status_code(req) == 401){
         err <- "Impossible to connect to GeoServer: Wrong credentials"
+        cli::cli_alert_danger(err)
         self$ERROR(err)
         stop(err)
       }
       if(status_code(req) == 404){
         err <- "Impossible to connect to GeoServer: Incorrect URL or GeoServer temporarily unavailable"
+        cli::cli_alert_danger(err)
         self$ERROR(err)
         stop(err)
       }
       if(status_code(req) != 200){
         err <- sprintf("Impossible to connect to Geoserver: Unexpected error (status code %s)", status_code(req))
+        cli::cli_alert_danger(err)
         self$ERROR(err)
         stop(err)
       }else{
-        self$INFO("Successfully connected to GeoServer!")
+        msg = "Successfully connected to GeoServer!"
+        cli::cli_alert_success(msg)
+        self$INFO(msg)
       }
       return(TRUE)
     },
@@ -232,7 +195,9 @@ GSManager <- R6Class("GSManager",
     #'@description Reloads the GeoServer catalog
     #'@return \code{TRUE} if reloaded, \code{FALSE} otherwise
     reload = function(){
-      self$INFO("Reloading GeoServer catalog")
+      msg = "Reloading GeoServer catalog"
+      cli::cli_alert_info(msg)
+      self$INFO(msg)
       reloaded <- FALSE
       req <- GSUtils$POST(self$getUrl(), private$user, 
                           private$keyring_backend$get(service = private$keyring_service, username = private$user), 
@@ -240,10 +205,14 @@ GSManager <- R6Class("GSManager",
                           content = NULL, contentType = "text/plain",
                           self$verbose.debug)
       if(status_code(req) == 200){
-        self$INFO("Successfully reloaded GeoServer catalog!")
+        msg = "Successfully reloaded GeoServer catalog!"
+        cli::cli_alert_success(msg)
+        self$INFO(msg)
         reloaded <- TRUE
       }else{
-        self$ERROR("Error while reloading the GeoServer catalog")
+        err = "Error while reloading the GeoServer catalog"
+        cli::cli_alert_danger(err)
+        self$ERROR(err)
       }
       return(reloaded)
     },
@@ -251,7 +220,9 @@ GSManager <- R6Class("GSManager",
     #'@description Get system status
     #'@return an object of class \code{data.frame} given the date time and metrics value
     getSystemStatus = function(){
-      self$INFO("Get system status")
+      msg = "Get system status"
+      cli::cli_alert_info(msg)
+      self$INFO(msg)
       datetime <- Sys.time()
       req <- GSUtils$GET(self$getUrl(), private$user, 
                          private$keyring_backend$get(service = private$keyring_service, username = private$user),
@@ -259,9 +230,13 @@ GSManager <- R6Class("GSManager",
                          contentType = "application/json",
                          self$verbose.debug)
       if(status_code(req) == 200){
-        self$INFO("Successfully fetched system status")
+        msg = "Successfully fetched system status"
+        cli::cli_alert_success(msg)
+        self$INFO(msg)
       }else{
-        self$ERROR("Error while fetching system status")
+        err = "Error while fetching system status"
+        cli::cli_alert_danger(err)
+        self$ERROR(err)
       }
       content <- httr::content(req)
       status <- list(