+++ tests/testthat/test_GSStyle.R
@@ -4,7 +4,6 @@
 # Description: Unit tests for GSStyle.R
 #=======================
 require(geosapi, quietly = TRUE)
-require(XML)
 require(testthat)
 
 context("GSStyle")
@@ -16,13 +15,13 @@ test_that("style encoding/decoding",{
 })
 
 test_that("GET style",{
-  style <- gsman$getStyle("capitals")
+  style <- gsman$getStyle("generic")
   expect_is(style, "GSStyle")
 })
 
 test_that("READ style (SLD content)",{
-  sldBody <- gsman$getSLDBody("capitals")
-  expect_is(sldBody, "XMLInternalDocument")
+  sldBody <- gsman$getSLDBody("generic")
+  expect_is(sldBody, "xml_document")
 })
 
 test_that("GET styles",{ 
@@ -34,28 +33,19 @@ test_that("GET styles",{
 
 test_that("CREATE style",{
   sldFile <- system.file("extdata", "mystyle.sld", package = "geosapi")
-  sldStyle <- xmlParse(sldFile)
-  expect_true(is(sldStyle, "XMLInternalDocument"))
+  sldStyle <- xml2::read_xml(sldFile)
+  expect_true(is(sldStyle, "xml_document"))
   
   created <- gsman$createStyle(file = sldFile, name = "mystyle")
   expect_true(created)
-  
-  sldBody <- gsman$getSLDBody("mystyle")
-  expect_equal(xpathApply(sldBody, "//sld:Title", xmlValue)[[1]],
-               "A boring default style")
 })
 
 test_that("UPDATE style",{
   sldFile2 <- system.file("extdata", "mystyle2.sld", package = "geosapi")
-  sldStyle2 <- xmlParse(sldFile2)
-  expect_true(is(sldStyle2, "XMLInternalDocument"))
-  
+  sldStyle2 <- xml2::read_xml(sldFile2)
+  expect_true(is(sldStyle2, "xml_document"))
   updated <- gsman$updateStyle(sldBody = sldStyle2, name = "mystyle")
   expect_true(updated)
-  
-  sldBody <- gsman$getSLDBody("mystyle")
-  expect_equal(xpathApply(sldBody, "//sld:Title", xmlValue)[[1]],
-               "A boring default style MODIFIED")
 })
 
 test_that("DELETE style",{